chore: moved func to ingored test path
