+++ packages/api-key-service/src/functions/__tests__/create.api-key.test.ts
@@ -12,7 +12,7 @@ jest.mock('@reapit/connect-session', () => ({
 import { createApiKey } from './../'
 import { HttpStatusCode } from '@homeservenow/serverless-aws-handler'
 import { Context } from 'aws-lambda'
-import { mockRequestHandlerContext } from './mock.hander.context'
+import { mockRequestHandlerContext } from '../tests/mock.hander.context'
 
 describe('Create ApiKey', () => {
   afterAll(() => {
